+++ spec/lib/ember_cli/runner_spec.rb
@@ -3,32 +3,69 @@ require "ember_cli/runner"
 describe EmberCli::Runner do
   describe "#run!" do
     context "when the command fails" do
-      it "writes STDERR and STDOUT to `err`" do
-        out = StringIO.new
-        err = StringIO.new
-        runner = EmberCli::Runner.new(err: err, out: out)
+      it "writes output to `out` streams" do
+        stdout = StringIO.new
+        logfile = StringIO.new
+        runner = EmberCli::Runner.new(
+          err: [],
+          out: [stdout, logfile],
+        )
 
-        expect { runner.run!("echo 'out'; echo 'err' > /dev/stderr; exit 1") }.
+        expect { runner.run!(command_with_error(out: "out")) }.
           to raise_error(SystemExit)
 
-        [err, out].each(&:rewind)
+        expect(split_output_from_stream(stdout)).to eq(%w[out out])
+        expect(split_output_from_stream(logfile)).to eq(%w[out out])
+      end
 
-        expect(err.read).to match(/out\nerr/)
-        expect(out.read).to eq("out\nerr\n")
+      it "writes errors to `err` streams" do
+        stderr = StringIO.new
+        logfile = StringIO.new
+        runner = EmberCli::Runner.new(
+          err: [stderr, logfile],
+          out: [],
+        )
+
+        expect { runner.run!(command_with_error(err: "err")) }.
+          to raise_error(SystemExit)
+
+        expect(split_output_from_stream(stderr)).to eq(%w[err err])
+        expect(split_output_from_stream(logfile)).to eq(%w[err err])
       end
     end
 
     it "executes the command" do
       out = StringIO.new
       err = StringIO.new
-      runner = EmberCli::Runner.new(err: err, out: out)
+      runner = EmberCli::Runner.new(err: [err], out: [out])
 
-      runner.run!("echo 'out'")
+      status = runner.run!(command("out"))
 
       [err, out].each(&:rewind)
 
-      expect(err.read).to be_empty
-      expect(out.read).to eq("out\n")
+      expect(status).to be_success
+      expect(split_output_from_stream(err)).to be_empty
+      expect(split_output_from_stream(out)).to eq(%w[out])
+    end
+
+    def split_output_from_stream(stream)
+      stream.rewind
+
+      stream.read.split
+    end
+
+    def command(output)
+      "echo '#{output}'"
+    end
+
+    def command_with_error(out: "", err: "")
+      [
+        "echo '#{out}'",
+        "echo '#{err}' > /dev/stderr",
+        "echo '#{out}'",
+        "echo '#{err}' > /dev/stderr",
+        "exit 1",
+      ].join("; ")
     end
   end
 end